+++ js/data.js
@@ -1,0 +1,53 @@
+'use strict'
+
+// Обьявляем константы для задания
+const INITIAL_POST_COUNT = 25; // по-умаолчанию надо 25, для теста поставил 2
+
+const DESCRIPTION_LIST = [
+  'Жить только для себя — есть злоупотребление.',
+  'Приятно следовать внушениям совести.',
+  'Глупость бывает двух родов: молчаливая и болтливая.',
+  'Для людей работа является наслаждением.',
+  'Честь дороже жизни.',
+  'Целью общества является всеобщее счастье.',
+  'Верная любовь помогает переносить все тяжести.',
+  'Кто стал дружить с тобой для обретенья благ, Не друг надежный твой, а самый страшный враг.',
+]
+
+const NAMES_LIST = [
+  'Сайвард "Беспощадный"',
+  'Элвина "Безумный Глаз"',
+  'Черноглазый Бизон',
+  'Бодрая Ваделси',
+  'Чудак Крoсли',
+  'Энор Чандлер',
+  'Кен Саймон',
+  'Крофтон Мерсер',
+  'Фэй Шортер',
+  'Адисон Фишоп',
+  'Симисола',
+  'Ифунания',
+  'Энитан',
+  'Абена',
+  'Анули',
+  'Тлексиктли',
+  'Зьянья',
+  'Найт',
+  'Ачкохтли',
+  'Ихуикатл',
+]
+
+const MESSAGES_LIST = [
+  'Всё отлично!',
+  'Подтверждено: солнечных дней всё меньше',
+  'В целом всё неплохо. Но не всё.',
+  'Оказывается, средства индивидуальной защиты оказались бесполезны!',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+  'Внезапно, герцог графства коронован',
+  'Противоположная точка зрения подразумевает, что сделанные на базе интернет-аналитики выводы разоблачены.',
+]
+
+export {INITIAL_POST_COUNT, DESCRIPTION_LIST, MESSAGES_LIST, NAMES_LIST};
